src/test/java/pl/ekulka/ecommerce/catalog/ProductCatalogTest.java
package pl.ekulka.ecommerce.catalog;

import org.junit.jupiter.api.Test;

import java.util.List;

import static org.junit.jupiter.api.Assertions.*;
import static org.assertj.core.api.Assertions.*;

public class ProductCatalogTest {

    @Test
    void itAllowsListingProducts() {
        ProductCatalog catalog = thereIsProductCatalog();

        List<Product> products = catalog.allProducts();

        assert products.isEmpty();
    }

    private ProductCatalog thereIsProductCatalog() {
        return new ProductCatalog();
    }

    @Test
    void itAllowsToAddProduct() {
        ProductCatalog catalog = thereIsProductCatalog();

        catalog.addProduct("Legoset 8083", "nice one");

        List<Product> allProducts = catalog.allProducts();

        assertThat(allProducts)
                .hasSize(1);
    }
}
